Adding main loop code for the updated msatcommander GUI
This will be the main loop of the updated msatcommander GUI.  It's been updated for the new GUI and the new msat and p3wrapr modules.  The current version is working for finding microsatellite repeats using the contents of the msat module.  I did all of this (stupidly) outside version control because I was interested in seeing how the pieces fit together.